treewide: s/format/seastar::format/
this change replace all `format()` with `seastar::format()`.

this was not necessary if we don't change the function signature of
`seastar::format()`, even if we compile the tree with libstdc++
shipped with GCC-13, which implements `std::format()`. because
`std::format()` accepts a templated parameter of `fmt`, so the
`seastar::fmt()` which accepts a plain `const char*` always win
if we just call `format()` in the opened namespace of `seastar`.

but since we plan to change `seastar::format()` from
`format(const char* fmt, A&&... a)` to
`format(fmt::format_string<A...> fmt, A&&... a)`, we cannot
gurantee that `seastar::format()` will win in the function
resolution in the use case above. so we have to change all
call sites of this function to `seastar::format()` to disambiguate
it.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>